requirements.txt
telethon>=1.35.0
pydantic>=2.6.0
PyYAML>=6.0
langdetect>=1.0.9
requests>=2.32.0
streamlit>=1.35.0
deepl>=1.9.0
pandas>=2.2.0
matplotlib>=3.10.6
wordcloud>=1.9.2
scikit-learn>=1.7.1
streamlit-autorefresh>=1.0.1
PySocks>=1.7.1
deep-translator==1.11.4
httpx>=0.27.0
SudachiPy>=0.6.10
sudachidict_core>=20250825